+++ src/main/java/insane96mcp/iguanatweaksreborn/integration/ToolBeltIntegration.java
@@ -1,5 +1,6 @@
-package insane96mcp.iguanatweaksreborn.module.sleeprespawn.death.integration;
+package insane96mcp.iguanatweaksreborn.integration;
 
+import dev.gigaherz.toolbelt.ToolBelt;
 import dev.gigaherz.toolbelt.belt.ToolBeltItem;
 import dev.gigaherz.toolbelt.slot.BeltExtensionSlot;
 import insane96mcp.iguanatweaksreborn.data.generator.ISOItemTagsProvider;
@@ -8,11 +9,12 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraftforge.common.util.LazyOptional;
 
 import java.util.List;
 
-public class ToolBelt {
+public class ToolBeltIntegration {
     public static final TagKey<Item> TOOL_BELT_TICKABLE = ISOItemTagsProvider.create("toolbelt_tickable");
     public static void onDeath(List<ItemStack> items, Player player) {
         LazyOptional<BeltExtensionSlot> oBeltExtensionSlot = BeltExtensionSlot.get(player);
@@ -31,12 +33,30 @@ public class ToolBelt {
         oBeltExtensionSlot.ifPresent(beltExtensionSlot -> {
             beltExtensionSlot.getSlots().forEach(extensionSlot -> {
                 ItemStack belt = extensionSlot.getContents();
-                if (belt.is(dev.gigaherz.toolbelt.ToolBelt.BELT.get())) {
+                if (belt.is(ToolBelt.BELT.get())) {
                     belt.getCapability(ToolBeltItem.ITEM_HANDLER).ifPresent(cap -> {
                         int slots = cap.getSlots();
                         for (int i = 0; i < slots; i++) {
                             if (cap.getStackInSlot(i).is(TOOL_BELT_TICKABLE))
                                 cap.getStackInSlot(i).inventoryTick(entity.level(), entity, -1, false);
+                        }
+                    });
+                }
+            });
+        });
+    }
+
+    public static void removeCursedItemsOnDeath(LivingEntity entity) {
+        LazyOptional<BeltExtensionSlot> oBeltExtensionSlot = BeltExtensionSlot.get(entity);
+        oBeltExtensionSlot.ifPresent(beltExtensionSlot -> {
+            beltExtensionSlot.getSlots().forEach(extensionSlot -> {
+                ItemStack belt = extensionSlot.getContents();
+                if (belt.is(ToolBelt.BELT.get())) {
+                    belt.getCapability(ToolBeltItem.ITEM_HANDLER).ifPresent(cap -> {
+                        int slots = cap.getSlots();
+                        for (int i = 0; i < slots; i++) {
+                            if (EnchantmentHelper.hasVanishingCurse(cap.getStackInSlot(i)))
+                                cap.extractItem(i, 1, false);
                         }
                     });
                 }