+++ src/main/java/insane96mcp/iguanatweaksreborn/module/items/ToolBeltIntegrationFeature.java
@@ -3,6 +3,7 @@ package insane96mcp.iguanatweaksreborn.module.items;
 import dev.gigaherz.toolbelt.ToolBelt;
 import dev.gigaherz.toolbelt.belt.ToolBeltItem;
 import insane96mcp.iguanatweaksreborn.InsaneSurvivalOverhaul;
+import insane96mcp.iguanatweaksreborn.integration.ToolBeltIntegration;
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.iguanatweaksreborn.module.misc.DataPacks;
 import insane96mcp.insanelib.base.Feature;
@@ -18,7 +19,7 @@ import net.minecraftforge.fml.ModList;
 
 @Label(name = "gigaherz Tool Belt")
 @LoadFeature(module = Modules.Ids.ITEMS)
-public class ToolBeltIntegration extends Feature {
+public class ToolBeltIntegrationFeature extends Feature {
 
 	@Config
 	@Label(name = "Bigger base belt", description = "Enables a data pack that changes the crafting of the Tool Belt to give more slots (2 -> 4)")
@@ -27,7 +28,7 @@ public class ToolBeltIntegration extends Feature {
 	@Label(name = "Reduce upgrade cost", description = "Reduces cost to apply pouches to tool belts")
 	public static Boolean reduceUpgradeCost = true;
 
-	public ToolBeltIntegration(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+	public ToolBeltIntegrationFeature(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
 		InsaneSurvivalOverhaul.addServerPack("toolbelt_integration", "Insane's Survival Overhaul Tool Belt Integration", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && biggerBaseBelt);
 	}
@@ -56,6 +57,6 @@ public class ToolBeltIntegration extends Feature {
 	public void onLivingTick(LivingEvent.LivingTickEvent event) {
 		if (!this.isEnabled())
 			return;
-		insane96mcp.iguanatweaksreborn.module.sleeprespawn.death.integration.ToolBelt.tryTickItemsIn(event.getEntity());
+		ToolBeltIntegration.tryTickItemsIn(event.getEntity());
 	}
 }